sdk/python/requirements.txt
pynacl>=1.5.0
msgpack>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
